refactored webhooks msg sending to make it configurable
